fixing bugs

+++ db/seeds.rb
@@ -7,12 +7,18 @@
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
 User.destroy_all
-Work.destroy_all
 
 User.create!(email: "albeiro63@yahoo.com.ar", password: "1234", password_confirmation: "1234", admin: true)
 
-image=File.open(File.join(Rails.root, "db/images/missing.png"))
+if Rails.env.development?
+	
+	Work.destroy_all
 
-50.times do |n|
-	Work.create!(name: Faker::Company.name, date: DateTime.now, description: Faker::Lorem.paragraph, image: image, price: Random.rand(10000..45000))
-end
+	image=File.open(File.join(Rails.root, "db/images/missing.png"))
+
+	50.times do |n|
+		Work.create!(name: Faker::Company.name, date: DateTime.now, description: Faker::Lorem.paragraph, image: image, price: Random.rand(10000..45000))
+	end
+	
+else
+
